requirements.txt
groq>=0.9.0,<1.0.0
tavily-python>=0.3.0,<1.0.0
playwright>=1.40.0,<2.0.0
playwright-stealth>=1.0.6,<2.0.0
trafilatura>=1.6.0,<2.0.0
transformers>=4.40.0,<=4.45.2
torch
optimum[onnxruntime]>=1.17.0,<=1.20.0
pandas>=2.0.0,<3.0.0
pyarrow>=14.0.0,<18.0.0
python-dotenv>=1.0.0,<2.0.0
rich>=13.0.0,<14.0.0
numpy<2.0.0

# Web Dashboard
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
sse-starlette>=2.0.0,<3.0.0
aiofiles>=23.2.0,<24.0.0
